refactor(providers): move rerank constants to models and centralize bedrock error parsing
